Strip the BOM without removing the leading 2 characters
If the string is a UTF-8 string then the 3 bytes that make up the BOM are
treated as a single character, meaning that when we do:

    css_string[3..-1]

We actually remove the BOM and the next 2 characters.  If we change this
to:

    css_string[1..-1]

then we correctly trim off only the BOM character.  However, trimming the
first character off only works if the string actually is encoded as UTF-8
so we force the string to UTF-8 before we trim it.  For example if we were
somehow given an ASCII encoding string that contained the BOM as the first
3 bytes these would be treated as the first 3 characters, so trimming the
first character only wouldn't correctly remove the BOM.  It's unlikely that
we'd be given an ASCII encoded string that started with the BOM characters
but if we were and didn't remove them it'd be written out to disk and then
treated as a UTF-8 with BOM file because it starts with those 3 bytes.

The downside to this is that we're modifying the supplied string by forcing
an encoding onto it before we manipulate it.  For our usecase this is
probably ok, as we only care about the string returned from the remove_bom
method, not the one passed in.  It'd be nice to work out an alternative if
we can though, as it's not super cool to change objects you don't return.