Performance improvements to removing the BOM
1. Replace the `string.bytes[0..2] == [0xEF, 0xBB, 0xBF]` check with
a string equality check on the first character of the string.  We need
to make sure we're dealing with a utf8 string to perform this check
correctly so we force the string to utf8 and do the check, but then set
it back to its old encoding afterwards.  `String#force_encoding` modifies
self so it would be bad-form if running remove_bom on an latin-1 encoded
string resulted in it being encoded in utf-8 afterwards.

2. Replace the `string.force_encoding('utf-8')[1..-1]` way of removing the
first character of the string with one that a) doesn't modify the original
string, and is faster (ironically by modifying the duplicate with
`String#slice!`.  As `String#slice!` returns the deleted character instead
of `self` we need to use `Object#tap` to make sure we return the modified
string and not the bit of it we've sliced out.

To support these changes we refactor the BomRemover to be a class instead
of a module and wrap it up in a class method invoker to keep the API the
same (e.g. `BomRemover.remove_bom(string)`).

Both of these improvements were found by implementing many alternate
implementations and running `Benchmark.bmbm` to find the fastest.  The
original method of checking for the presence of the utf-8 string was
much slower than others, and while we don't get a huge speed boost from
our new bom removal method it is safer and still faster than some
alternatives.

See: https://gist.github.com/h-lame/e5ec2e5dcfa58abaccb693c28ed28f45
for details of the benchmarking and alternate implementations we tried.